monigrid-be/requirements.txt
# Python 3.13+
# JAVA_HOME 환경변수 필요 (JDK 11+)

# Web framework
Flask==3.1.3
flask-cors==6.0.2
flask-limiter==4.1.1

# Production WSGI server (Windows-friendly, used by Windows service deployment)
waitress==3.0.2

# Validation & security
pydantic==2.12.5
PyJWT==2.12.1
python-dotenv==1.2.2
bcrypt==4.2.0

# JDBC bridge (JayDeBeApi depends on JPype1)
# JPype1 1.6.0 segfaults on Windows + Python 3.12 (access violation on `import jpype`).
# Pin to 1.5.2 until the upstream fix lands.
JayDeBeApi==1.2.3
JPype1==1.5.2

# SSH remote command execution (server resource monitoring)
paramiko==4.0.0

# WinRM remote command execution (Windows server resource monitoring)
pywinrm==0.4.3

# HTTP client (health-check proxy)
requests==2.33.1

# EXE packaging
pyinstaller==6.19.0

